+++ src/main/java/com/mca/repository/PromotionRepository.java
@@ -8,5 +8,5 @@ import com.mca.infrastructure.model.Promotion;
 import java.util.List;
 
 public interface PromotionRepository extends JpaRepository<Promotion, Long> {
-     List<Object[]> findVideoGameIdAndPromotionByVideoGameId(@Param("videoGameId") Integer videoGameId);
+     List<Object[]> findVideoGameIdAndPromotionByVideoGameId(@Param("videoGameId") Long videoGameId);
 }
